Changed default paging size
The default paging size was decreased from 500 to 100 because increasing the group size increased query time on solr. Patients with follow ups where stripped, that is why the  group.limit was increased from 100 to 1000. In the future maybe a different structure needs to be indexed, instead of each series as a single file, a parent child relationship between patient and study would be benefical.